add point cloud request
